+++ src/main/java/de/b0n/dir/model/DupFinderModel.java
@@ -1,18 +1,12 @@
 package de.b0n.dir.model;
 
 import com.github.funthomas424242.unmodifiable.UnmodifiableQueue;
-import com.github.funthomas424242.unmodifiable.UnmodifiableQueueFIFO;
-import de.b0n.dir.processor.AbstractProcessorModel;
-
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.Queue;
-import java.util.concurrent.ConcurrentLinkedQueue;
+import de.b0n.dir.processor.SearchProcessorModel;
 
 /**
  * Created by huluvu424242 on 05.03.17.
  */
-public class DupFinderModel<G,E> extends AbstractProcessorModel<G,E> {
+public class DupFinderModel<G,E> extends SearchProcessorModel<G,E> {
 
     public DupFinderModel(){
 
@@ -24,7 +18,7 @@ public class DupFinderModel<G,E> extends AbstractProcessorModel<G,E> {
         return new ModelFactory(){
 
             @Override
-            public AbstractProcessorModel createModel() {
+            public SearchProcessorModel createModel() {
                 return new DupFinderModel();
             }
         };
@@ -32,10 +26,10 @@ public class DupFinderModel<G,E> extends AbstractProcessorModel<G,E> {
 
     /**
      * Liefert die Gesamtanzahl der Elemente über alle Gruppen. Falls dieser
-     * AbstractProcessorModel mehr als Integer.MAX_VALUE Elemente enthält, liefert er
+     * SearchProcessorModel mehr als Integer.MAX_VALUE Elemente enthält, liefert er
      * Integer.MAX_VALUE.
      *
-     * @return Anzahl aller Elemente im AbstractProcessorModel
+     * @return Anzahl aller Elemente im SearchProcessorModel
      */
     public int size() {
         int size = 0;
