+++ src/main/java/de/b0n/dir/processor/SearchProcessor.java
@@ -5,7 +5,7 @@ import de.b0n.dir.DupFinderCallback;
 /**
  * Created by huluvu424242 on 25.02.17.
  */
-public abstract class AbstractSearchProcessor {
+public abstract class SearchProcessor {
 
     protected static final DupFinderCallback DUMMY_CALLBACK = new DupFinderCallback(){};
